feat(scoring): muda stageMultiplier de Int para Float (1x, 1.5x, 2x, 2.5x) com ROUND no trigger
- Supabase: migration altera coluna stage_multiplier para DECIMAL(3,1) e atualiza trigger com ROUND()
- Match.kt / MatchDto.kt: tipo Int -> Float
- PredictionCalculator: parametros Float + roundToInt() para pontos sempre inteiros
- PredictionCalculatorTest: atualiza casos existentes e adiciona testes para 1.5x, 2.0x e 2.5x
- MatchPredictionCard: novo componente StageMultiplierBadge exibido no header quando multiplicador > 1.0x
- README: tabela de fases com novos multiplicadores

+++ shared/src/commonMain/kotlin/com/bolao/domain/model/Match.kt
@@ -29,7 +29,7 @@ data class Match(
     val homeOdd: Double? = null,
     val drawOdd: Double? = null,
     val awayOdd: Double? = null,
-    val stageMultiplier: Int = 1,
+    val stageMultiplier: Float = 1.0f,
     val status: GameStatus = GameStatus.Scheduled,
     val scheduledAt: Instant,
     val competition: String,
